requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
lxml==6.0.0
MarkupSafe==3.0.2
numpy==2.3.1
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.24.10
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
pytz==2025.2
RapidFuzz==3.13.0
scikit-learn==1.7.1
scipy==1.16.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
pytest 
httpx